Switch upd_proxy id hash function to 32 bit one(!!!)
You need to delete the cnc-msl/msl_upd_proxy/proxy_gen in order to
regenerate the appropriete source code.

+++ alica-essentials/udp_proxy_generator/include/RelayedMessage.h
@@ -5,6 +5,7 @@
 #include <functional>
 
 #include <boost/regex.hpp>
+#include <cstdint>
 
 using namespace std;
 
@@ -26,7 +27,7 @@ public:
 	int Udp2RosQueueLength;
 
 	std::hash<std::string> hash;
-	size_t Id;
+	uint32_t Id;
 
 	string getRosCallBackName();
 	string getRosClassName();
